--- aktual-budget/data/db/src/androidHostTest/jniLibs/.gitignore
@@ -1,4 +1,0 @@
-# Native libraries extracted from sqlite-bundled-jvm at build time — not committed
-*.so
-*.dylib
-*.dll
